requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.0.8
django-ckeditor==5.9.0
django-filter==2.3.0
django-js-asset==1.2.2
django-pure-pagination==0.3.0
django-rest-multiple-models==2.1.3
django-shortuuidfield==0.1.3
django-simpleui==4.0.3
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.5.3
PyMySQL==0.10.0
python-memcached==1.59
pytz==2020.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
